Rename header folder to headers
